refactor: 디테일 페이지 구조화 시키기
